% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Data_Descriptions.R
\docType{data}
\name{author_attributes}
\alias{author_attributes}
\title{Department Manager Email dataset -- Author Attributes}
\format{A dataframe with 20 rows and 2 columns.}
\usage{
data(author_attributes)
}
\description{
A dataset containing 121 emails sent between department managers in a North
Carolina County Government. This dataset is very small and not terrbily interesting
and is meant primarily to serve as an illustration of what the model can do and the
output it will create. The data objects included represent the minimal set necessary
to make themodel work and should serve as a template for formatting your own data
and are as follows:
}
\details{
\itemize{
  \item Department -- The name of the department the individual is a manager of.
  \item Gender -- The gender of the department manager.
}
}
\keyword{datasets}

