% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Cluster_Run_Models.R
\name{Cluster_Run_Models}
\alias{Cluster_Run_Models}
\title{A Function to run the ContentStructure model to convergence for multiple datasets in parallel.}
\usage{
Cluster_Run_Models(name_vector, Data_Directory, MAX_CPUs = 4,
  Main_Iterations = 4000, Sample_Step_Burnin = 2e+06,
  Sample_Step_Iterations = 8e+06, Topics = 40, Clusters = 4,
  Sample_Step_Sample_Every = 2000)
}
\arguments{
\item{name_vector}{A vector containing the exact names of the .Rdata files containing the relevant files necessary to run the model for each organization.}

\item{Data_Directory}{The directory where all .Rdata files are stored -- also where all output will be stored.}

\item{MAX_CPUs}{The maximum number of datasets that can be analyzed in parallel -- should be less than or equal to the number of cores on your machine.}

\item{Main_Iterations}{The number of iterations of Gibbs sampling for the LDA part of the model. We have found 4,000 seems to work well.}

\item{Sample_Step_Burnin}{The number of iterations of burnin that should be completed before sampling the latent space parameters when running MH for the LSM to convergence.}

\item{Sample_Step_Iterations}{The total number of iterations to run MH for the LSM for (before thinning).}

\item{Topics}{The number of topics to use}

\item{Clusters}{The number of topic clusters to use}

\item{Sample_Step_Sample_Every}{How many iterations to skip when thinning the MH for the LSM chain in our MH for the LSM sample step.}
}
\value{
Does not return anything, just saves everything to our data_directory folder.
}
\description{
A Function to run the ContentStructure model to convergence for multiple datasets in parallel.
}

