% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Data_Descriptions.R
\docType{data}
\name{document_edge_matrix}
\alias{document_edge_matrix}
\title{Department Manager Email dataset -- Document Edge Matrix}
\format{A a number of emails (121) by number of authors plus one (21), dataframe where the first column records the index (starting from 1) of the sender of the email and the rest of the columns record a 1 if the manager recieved the email and zero otherwise.}
\usage{
data(document_edge_matrix)
}
\description{
A dataset containing 121 emails sent between department managers in a North
Carolina County Government. This dataset is very small and not terrbily interesting
and is meant primarily to serve as an illustration of what the model can do and the
output it will create. The data objects included represent the minimal set necessary
to make themodel work and should serve as a template for formatting your own data
and are as follows:
}
\details{
\itemize{
  \item The first column records the index (starting from 1) of the sender of the email.
  \item The remaining 20 columns  (1 for each manager in the same order they appear in the author_attributes dataset) are equal to 1 if the manager recieved (or was Cc'd on) the email and zero otherwise.
  \item There is one row for each email.
}
}
\keyword{datasets}

