% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Data_Descriptions.R
\docType{data}
\name{document_word_matrix}
\alias{document_word_matrix}
\title{Department Manager Email dataset -- Document Word Matrix}
\format{A a number of emails (121) by number of unique terms (204) dataframe that records the number of times each term was used in each email.}
\usage{
data(document_word_matrix)
}
\description{
A dataset containing 121 emails sent between department managers in a North
Carolina County Government. This dataset is very small and not terrbily interesting
and is meant primarily to serve as an illustration of what the model can do and the
output it will create. The data objects included represent the minimal set necessary
to make themodel work and should serve as a template for formatting your own data
and are as follows:
}
\details{
\itemize{
  \item document_word_matrix -- a number of emails (121) by number of unique terms (204) matrix the records the number of times each term was used in each email. Note that stopwords
  have been removed as have names.
}
}
\keyword{datasets}

