% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Generate_Diagnostics_All_Datasets.R
\name{Generate_Diagnostics_All_Datasets}
\alias{Generate_Diagnostics_All_Datasets}
\title{A Function to generate diagnostic plots and interpretable output for a list of 1 to N model output objects from running the ContentStructure model.}
\usage{
Generate_Diagnostics_All_Datasets(data_list, only_generate_summaries = T,
  data_directory, print_agg_stats = F, using_county_email_data = F)
}
\arguments{
\item{data_list}{A vector containing the names of organization data files as in the function to run the model. Should only contain names of those organizations for which output has already been created.}

\item{only_generate_summaries}{If TRUE, then only generate a one-page-per-cluster pdf summary of model output for each county, otherwise generate a ton of output.}

\item{data_directory}{The directory where all .Rdata files are stored -- also where all output will be stored.}

\item{print_agg_stats}{If TRUE, then will print aggregate level stats for the county across all cluster simultaneously. I have not found this to be terribly useful.}

\item{using_county_email_data}{Logical if you are using North Carolina County Government email data that are properly formatted to produce aggregate level output.}
}
\value{
A list object with the following structure. The first entry is a sublist which contains three objects: a vector containing the full vocabulary, an email-word matrix aggregated across all organizations, and a metadata data frame -- again aggregated across all emails in all organizations. The second entry in the outer list is a list with one entry for each organization with a sublist containing data about that organization and emails sent within it. Only returns if you are using North Carolina County Government Email data.
}
\description{
A Function to generate diagnostic plots and interpretable output for a list of 1 to N model output objects from running the ContentStructure model.
}

