% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Generate_Heldout_Samples.R
\name{Generate_Heldout_Samples}
\alias{Generate_Heldout_Samples}
\title{A Function to generate test-train splits using NH county data. This is meant for internal use with our own data format.}
\usage{
Generate_Heldout_Samples(files, seed, held_out_proportion = 0.1)
}
\arguments{
\item{files}{A string vector containing the names of the well-formatted .Rdata files with data from the counties we wish to generate a test train split from.}

\item{seed}{The seed we wish to use to generate the random splits. This is set up so we can vary it and so we can reproduce our results.}

\item{held_out_proportion}{A number between zero and 1, defaults to 0.1}
}
\value{
Returns a list of lists where each sub-list contains a test-train split of the data generated by the chosen seed.
}
\description{
A Function to generate test-train splits using NH county data. This is meant for internal use with our own data format.
}

