% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Data_Descriptions.R
\docType{data}
\name{vocabulary}
\alias{vocabulary}
\title{Department Manager Email dataset -- Vocabulary}
\format{A dataframe with 20 rows and 2 columns.}
\usage{
data(vocabulary)
}
\description{
A dataset containing 121 emails sent between department managers in a North
Carolina County Government. This dataset is very small and not terrbily interesting
and is meant primarily to serve as an illustration of what the model can do and the
output it will create. The data objects included represent the minimal set necessary
to make themodel work and should serve as a template for formatting your own data
and are as follows:
}
\details{
\itemize{
  \item a vector containing all unique terms.
}
}
\keyword{datasets}

