% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Run_Full_Model.R
\name{Run_Full_Model}
\alias{Run_Full_Model}
\title{A Function to run the ContentStructure model to convergence for one dataset.}
\usage{
Run_Full_Model(data_name, main_iterations = 4000,
  sample_step_burnin = 2e+06, sample_step_iterations = 8e+06,
  sample_step_sample_every = 2000, topics = 40, clusters = 4,
  data_directory, run_MH_only = F)
}
\arguments{
\item{data_name}{The exact name of the .Rdata file containing the relevant files necessary to run the model.}

\item{main_iterations}{The number of iterations of Gibbs sampling for the LDA part of the model. We have found 4,000 seems to work well.}

\item{sample_step_burnin}{The number of iterations of burnin that should be completed before sampling the latent space parameters when running MH for the LSM to convergence.}

\item{sample_step_iterations}{The total number of iterations to run MH for the LSM for (before thinning).}

\item{sample_step_sample_every}{How many iterations to skip when thinning the MH for the LSM chain in our MH for the LSM sample step.}

\item{topics}{The number of topics to use}

\item{clusters}{The number of topic clusters to use}

\item{data_directory}{The directory where our data_name file is stored. This is also where all output will be saved}

\item{run_MH_only}{If TRUE, then we only rerun MH for the LSM to convergence}
}
\value{
Does not return anything, just saves everything to our data_directory folder.
}
\description{
A Function to run the ContentStructure model to convergence for one dataset.
}

